Moved block compression to a module
